+++ examples/manual_option/src/main.rs
@@ -1,15 +1,13 @@
 use chrono::{Datelike, TimeZone, Timelike};
 use chrono_tz::UTC;
-use rrule::{Frequency, Options, RRule};
+use rrule::{Frequency, ParsedOptions, RRule};
 
 fn main() {
     // Build options that starts first day in 2020 at 9:00AM and occurs daily 5 times
-    let options = Options::default()
+    let options = ParsedOptions::default()
         .dt_start(UTC.ymd(2020, 1, 1).and_hms(9, 0, 0))
         .count(5)
-        .freq(Frequency::Daily)
-        .build()
-        .unwrap();
+        .freq(Frequency::Daily);
 
     // Construct `RRule` from options
     let rrule = RRule::new(options).expect("RRule invalid");