+++ rrule/src/parser/mod.rs
@@ -1,12 +1,3 @@
-mod parse_options;
 mod rrulestr;
 
-pub use parse_options::parse_options;
-pub(crate) use rrulestr::{build_rruleset, parse_rrule_string_to_options};
-
-fn is_some_and_not_empty<T>(v: &Option<Vec<T>>) -> bool {
-    match v {
-        Some(v) => !v.is_empty(),
-        None => false,
-    }
-}
+pub(crate) use rrulestr::{build_rruleset, finalize_parsed_options, parse_rrule_string_to_options};